* Ensure plugin version is updated when a transform is re-run using a newer   plugin -- so that the plugin version saved in the session corresponds with   the one that was actually used, not the one previously specified * Fix uninitialised font size in preferences -- was causing all sorts of grief

+++ base/Preferences.cpp
@@ -62,7 +62,8 @@ Preferences::Preferences() :
     m_backgroundMode = BackgroundMode
         (settings.value("background-mode", int(BackgroundFromTheme)).toInt());
     m_viewFontSize = settings.value
-        ("view-font-size", QApplication::font().pointSize() * 0.9).toInt();
+        ("view-font-size", int(QApplication::font().pointSize() * 0.9))
+        .toInt();
     settings.endGroup();
 
     settings.beginGroup("TempDirectory");
